+++ apps/expo/src/app/_layout.tsx
@@ -1,19 +1,19 @@
-import { useEffect, useRef, useState } from "react";
-import { Platform, Text, View } from "react-native";
+import { Text, View } from "react-native";
 import { MenuProvider } from "react-native-popup-menu";
 import {
   SafeAreaProvider,
   useSafeAreaInsets,
 } from "react-native-safe-area-context";
-import Constants from "expo-constants";
-import * as Device from "expo-device";
-import * as Notifications from "expo-notifications";
-import { router, Stack } from "expo-router";
+import { Stack } from "expo-router";
 import * as SecureStore from "expo-secure-store";
 import { StatusBar } from "expo-status-bar";
 import { ClerkProvider, SignedIn } from "@clerk/clerk-expo";
 import { useColorScheme } from "nativewind";
 
+import {
+  NotificationProvider,
+  useNotification,
+} from "~/providers/NotificationProvider";
 import { TRPCProvider } from "~/utils/api";
 
 import "../styles.css";
@@ -34,10 +34,9 @@ export function ErrorBoundary({ error, retry }: ErrorBoundaryProps) {
 
 const tokenCache = {
   getToken: async (key: string) => {
-    const token = await SecureStore.getItemAsync(key, {
+    return await SecureStore.getItemAsync(key, {
       keychainAccessGroup: "group.soonlist.soonlist",
     });
-    return token;
   },
   saveToken: (key: string, value: string) => {
     return SecureStore.setItemAsync(key, value, {
@@ -46,183 +45,7 @@ const tokenCache = {
   },
 };
 
-Notifications.setNotificationHandler({
-  handleNotification: async () => ({
-    shouldShowAlert: true,
-    shouldPlaySound: false,
-    shouldSetBadge: false,
-  }),
-});
-
-export async function sendPushNotification(expoPushToken: string) {
-  const message = {
-    to: expoPushToken,
-    sound: "default",
-    title: "Want in early?",
-    body: "You can be one of the first. We want to learn together.",
-    data: { someData: "goes here" },
-  };
-
-  await fetch("https://exp.host/--/api/v2/push/send", {
-    method: "POST",
-    headers: {
-      Accept: "application/json",
-      "Accept-encoding": "gzip, deflate",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(message),
-  });
-}
-
-function handleRegistrationError(errorMessage: string) {
-  console.error(errorMessage);
-  throw new Error(errorMessage);
-}
-
-async function registerForPushNotificationsAsync() {
-  if (Platform.OS === "android") {
-    void Notifications.setNotificationChannelAsync("default", {
-      name: "default",
-      importance: Notifications.AndroidImportance.MAX,
-      vibrationPattern: [0, 250, 250, 250],
-      lightColor: "#FF231F7C",
-    });
-  }
-
-  if (Device.isDevice) {
-    const { status: existingStatus } =
-      await Notifications.getPermissionsAsync();
-    let finalStatus = existingStatus;
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-enum-comparison
-    if (existingStatus !== "granted") {
-      const { status } = await Notifications.requestPermissionsAsync();
-      finalStatus = status;
-    }
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-enum-comparison
-    if (finalStatus !== "granted") {
-      handleRegistrationError(
-        "Permission not granted to get push token for push notification!",
-      );
-      return;
-    }
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const projectId: string | undefined =
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      Constants.expoConfig?.extra?.eas?.projectId ??
-      Constants.easConfig?.projectId;
-    if (!projectId) {
-      handleRegistrationError("Project ID not found");
-    }
-    try {
-      const pushTokenString = (
-        await Notifications.getExpoPushTokenAsync({
-          projectId,
-        })
-      ).data;
-      console.log(pushTokenString);
-      return pushTokenString;
-    } catch (e: unknown) {
-      // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-      handleRegistrationError(`${e}`);
-    }
-  } else {
-    handleRegistrationError("Must use physical device for push notifications");
-  }
-}
-
-// function PushNotificationSenderButton({
-//   expoPushToken,
-// }: {
-//   expoPushToken: string;
-// }) {
-//   const notificationMutation =
-//     api.notification.sendSingleNotification.useMutation({});
-
-//   return (
-//     <Button
-//       title="Press to Send Notification"
-//       onPress={() => {
-//         notificationMutation.mutate({
-//           expoPushToken: expoPushToken,
-//           title: "Test from TRPC",
-//           body: "Test",
-//           data: { test: "test" },
-//         });
-//       }}
-//     />
-//   );
-// }
-
-function useNotificationObserver() {
-  useEffect(() => {
-    let isMounted = true;
-
-    function redirect(notification: Notifications.Notification) {
-      const url = notification.request.content.data.url as string | undefined;
-      if (url) {
-        router.push(url);
-      }
-    }
-
-    void Notifications.getLastNotificationResponseAsync().then((response) => {
-      if (!isMounted || !response?.notification) {
-        return;
-      }
-      redirect(response.notification);
-    });
-
-    const subscription = Notifications.addNotificationResponseReceivedListener(
-      (response) => {
-        redirect(response.notification);
-      },
-    );
-
-    return () => {
-      isMounted = false;
-      subscription.remove();
-    };
-  }, []);
-}
-
-// This is the main layout of the app
-// It wraps your pages with the providers they need
 export default function RootLayout() {
-  const [expoPushToken, setExpoPushToken] = useState("");
-  const insets = useSafeAreaInsets();
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [notification, setNotification] = useState<
-    Notifications.Notification | undefined
-  >(undefined);
-  const notificationListener = useRef<Notifications.Subscription>();
-  const responseListener = useRef<Notifications.Subscription>();
-  useNotificationObserver();
-
-  useEffect(() => {
-    registerForPushNotificationsAsync()
-      .then((token) => setExpoPushToken(token ?? ""))
-      .catch((error: string) => setExpoPushToken(`${error}`));
-
-    notificationListener.current =
-      Notifications.addNotificationReceivedListener((notification) => {
-        setNotification(notification);
-      });
-
-    responseListener.current =
-      Notifications.addNotificationResponseReceivedListener((response) => {
-        console.log(response);
-      });
-
-    return () => {
-      notificationListener.current &&
-        Notifications.removeNotificationSubscription(
-          notificationListener.current,
-        );
-      responseListener.current &&
-        Notifications.removeNotificationSubscription(responseListener.current);
-    };
-  }, []);
-
-  const { colorScheme } = useColorScheme();
   const clerkPublishableKey = process.env.EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY;
 
   if (!clerkPublishableKey) {
@@ -236,31 +59,42 @@ export default function RootLayout() {
   return (
     <ClerkProvider publishableKey={clerkPublishableKey} tokenCache={tokenCache}>
       <TRPCProvider>
-        <SafeAreaProvider>
-          <MenuProvider>
-            <View style={{ flex: 1 }}>
-              <Stack
-                screenOptions={{
-                  headerStyle: {
-                    backgroundColor: "#E0D9FF",
-                  },
-                  headerTintColor: "#5A32FB",
-                  contentStyle: {
-                    backgroundColor:
-                      colorScheme == "dark" ? "#09090B" : "#FFFFFF",
-                  },
-                }}
-              />
-              <AuthAndTokenSync expoPushToken={expoPushToken} />
-              <SignedIn>
-                <View style={{ paddingBottom: insets.bottom + 36 }} />
-                <BottomBar expoPushToken={expoPushToken} />
-              </SignedIn>
-            </View>
-            <StatusBar />
-          </MenuProvider>
-        </SafeAreaProvider>
+        <NotificationProvider>
+          <SafeAreaProvider>
+            <MenuProvider>
+              <RootLayoutContent />
+            </MenuProvider>
+          </SafeAreaProvider>
+        </NotificationProvider>
       </TRPCProvider>
     </ClerkProvider>
+  );
+}
+
+function RootLayoutContent() {
+  const insets = useSafeAreaInsets();
+  const { colorScheme } = useColorScheme();
+  const { expoPushToken } = useNotification();
+
+  return (
+    <View style={{ flex: 1 }}>
+      <Stack
+        screenOptions={{
+          headerStyle: {
+            backgroundColor: "#E0D9FF",
+          },
+          headerTintColor: "#5A32FB",
+          contentStyle: {
+            backgroundColor: colorScheme == "dark" ? "#09090B" : "#FFFFFF",
+          },
+        }}
+      />
+      <AuthAndTokenSync expoPushToken={expoPushToken} />
+      <SignedIn>
+        <View style={{ paddingBottom: insets.bottom + 36 }} />
+        <BottomBar expoPushToken={expoPushToken} />
+      </SignedIn>
+      <StatusBar />
+    </View>
   );
 }
